Remove old boulder, corla audit creation. continue to replace BallotPool.
