+++ core/src/testFixtures/kotlin/org/cryptobiotic/rlauxe/raire/SimulateRaireTestContest.kt
@@ -17,7 +17,7 @@ import org.cryptobiotic.rlauxe.util.df
 import org.cryptobiotic.rlauxe.util.listToMap
 import kotlin.random.Random
 
-// Simulation of Raire Contest; pass in the parameters and simulate the cvrs; then vall raire library to generate the assertions
+// Simulation of Raire Contest; pass in the parameters and simulate the cvrs; then call raire library to generate the assertions
 fun simulateRaireTestContest(N: Int, contestId: Int, ncands:Int, minMargin: Double, undervotePct: Double = .05, phantomPct: Double = .005, quiet: Boolean = true)
 : Pair<RaireContestUnderAudit, List<Cvr>> {
 
@@ -116,7 +116,15 @@ private fun trytoMakeRaireContest(N: Int, contestId: Int, ncands:Int, minMargin:
     return Pair(rcontestUA, testCvrs)
 }
 
-// TODO use SimulateRaireTestData
+data class SimulateIrvTestData(
+    val contest: RaireContest,
+    val minMargin: Double,
+    val sampleLimits: Int,
+    val excessVotes: Int? = null,
+    val quiet: Boolean = true
+)
+
+// TODO use SimulateIrvTestData
 data class RaireContestTestData(
     val contestId: Int,
     val ncands: Int,
