+++ plots/src/test/kotlin/org/cryptobiotic/rlauxe/workflow/SFoaSingleRoundAuditTaskGenerator.kt
@@ -7,10 +7,6 @@ import org.cryptobiotic.rlauxe.core.TestH0Result
 import org.cryptobiotic.rlauxe.estimate.ConcurrentTaskG
 import org.cryptobiotic.rlauxe.persist.PersistentAudit
 import org.cryptobiotic.rlauxe.persist.csv.AuditableCardCsvReaderSkip
-import org.cryptobiotic.rlauxe.persist.csv.readBallotPoolCsvFile
-import org.cryptobiotic.rlauxe.persist.csv.poolNameToId
-import org.cryptobiotic.rlauxe.util.SortMerge
-import org.cryptobiotic.rlauxe.util.createZipFile
 
 val skipPerRun = 8_000
